requirements1.txt

aiohttp==3.8.1
aiosignal==1.2.0
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.67
Babel==2.10.1
botocore==1.27.66
Brotli==1.0.9
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==37.0.2
cssselect2==0.6.0
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-admin-interface==0.19.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-form==3.4
django-bootstrap-modal-forms==2.2.0
django-click==2.3.0
django-colorfield==0.6.3
django-crispy-forms==1.14.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-htmx==1.12.1
django-mathfilters==1.0.0
django-pg-copy==1.0.0
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-weasyprint==2.1.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.13.1
docutils==0.16
et-xmlfile==1.1.0
firebase-admin==5.2.0
Flask==2.1.1
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
item==0.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
lxml==4.8.0
MarkupSafe==2.1.1
msgpack==1.0.3
multidict==6.0.2
numpy==1.23.2
openpyxl==3.0.10
oscrypto==1.3.0
pandas==1.4.3
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF3==1.0.6
pyphen==0.12.0
pyTelegramBotAPI==4.5.1
python-barcode==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytils==0.4.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
rsa==4.7.2
s3transfer==0.6.0
selection==0.0.14
self==2020.12.3
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
user-agent==0.1.10
weasyprint==54.3
webencodings==0.5.1
weblib==0.1.30
Werkzeug==2.1.1
whitenoise==6.0.0
xhtml2pdf==0.2.7
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.7.2
zipp==3.8.0
zopfli==0.2.1
